requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
mlxtend
# Para el chatbot
python-dotenv
google-generativeai